Refact quiz.js introudced local storage for title
