+++ index.html
@@ -75,7 +75,7 @@
                 This game is going to test your general knowledge of all things
                 from zodiac signs to video game release dates!
               </p>
-              <a href="./trivia_game.html" class="btn btn-primary"
+              <a href="./trivia_game.html" class="btn btn-primary" id="category_general_knowledge"
                 >Ready to play? Let's Go!</a
               >
             </div>
@@ -95,7 +95,7 @@
                 This game is going to test your knowledge of topics from medical
                 procedures to elements in the periodic table.
               </p>
-              <a href="./trivia_game.html" class="btn btn-primary"
+              <a href="./trivia_game.html" class="btn btn-primary" id="category_science_nature"
                 >Ready to play? Let's Go!</a
               >
             </div>
@@ -115,7 +115,7 @@
                 Think you know art like the back of your hand? Take this quiz
                 out for a spin!
               </p>
-              <a href="./trivia_game.html" class="btn btn-primary"
+              <a href="./trivia_game.html" class="btn btn-primary" id="category_art"
                 >Ready to play? Let's Go!</a
               >
             </div>
@@ -135,7 +135,7 @@
                 Lions, and tigers, and bears - oh my! Give this trivia game a
                 try!
               </p>
-              <a href="./trivia_game.html" class="btn btn-primary"
+              <a href="./trivia_game.html" class="btn btn-primary" id="category_animals"
                 >Ready to play? Let's Go!</a
               >
             </div>
@@ -152,8 +152,7 @@
         integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
         crossorigin="anonymous"
       ></script>
-      <script src="./layout.js"></script>
-      <script src="./script.js"></script>
+      <script type="module" src="./script.js"></script>
     </div>
   </body>
 </html>
