+++ trivia_game.html
@@ -28,6 +28,7 @@
         </li>
       </ul>
       <div id="alert_placeholder"></div>
+      <div><h1 id="quiz_title">Science</h1></div>
 
       <div id="question_container">
         <b>Question(s):</b>
@@ -60,10 +61,9 @@
                 aria-label="Close"
               ></button>
             </div>
-            <div class="modal-body" id="gifBody" >
+            <div class="modal-body" id="gifBody">
               <div id="gifs_container"></div>
               <b id="modal_text"></b>
-
             </div>
             <div class="modal-footer">
               <a href="./index.html"
@@ -88,8 +88,6 @@
     ></script>
     <script src="https://cdn.jsdelivr.net/npm/underscore@1.13.6/underscore-umd-min.js
     "></script>
-
-    <script type="module" src="./script.js"></script>
     <script type="module" src="./quiz.js"></script>
   </body>
 </html>
